Temporarily fix 'New Topic' button when using bbPress and BuddyPress.
bbPress uses a different anchor for the 'New Topic' button than BuddyPress.

The right approach would be to filter 'bp_get_group_new_topic_button' in
BuddyPress, but that hook didn't work properly. See:
https://buddypress.trac.wordpress.org/changeset/6534.

So we have to resort to removing BP's 'New Topic' button entirely and
creating a brand-new 'New Topic' button for now.

Fixes cuny-academic-commons/commons-in-a-box#34.