Fix 'New Topic' button for group forums.
bp_is_group_forum() changed in BP 1.7 to check if the legacy forums are
are explicitly installed.  This change broke the visibility of the
button when using the bbPress plugin.

In this commit, we use a customized version of bp_is_group_forum() to
do our group forum check.  Doing this ensures that the 'New Topic'
button will be displayed again when on group forum page.

Fixes #147.