Refactors automatic homepage generation to occur only on main site.
The CBOX Theme automatically generates a homepage and sets it as the front
page.  When using the theme on secondary sites on multisite, this isn't desirable if one simply wants to have the latest posts displayed.

This commit restricts homepage generation to the main BuddyPress site only.

Also, this commit adds some logic to determine if the homepage has been
accidentally removed.  If so, the frontpage DB options are deleted to
prevent a 404 on the front page.

Fixes #120.