Change the button ID on registration page from 'submit' to 'wp-submit'.
In commit 26c2af9, the button submit ID was changed from
'signup_submit' to 'submit' to allow the WP password strength script
to function properly.

However, changing the ID to 'submit' can conflict with JS apps. For
instance, reCAPTCHA cannot properly function because of this ID. See:
https://stackoverflow.com/a/42860847.

To workaround this, we will change the button ID to 'wp-submit', which
is also supported by the WP password strength meter JS:
https://github.com/WordPress/WordPress/blob/e82a76b282c02121bf0f970cde96a1f523f9fac7/wp-admin/js/user-profile.js#L106

Fixes #269.